% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{HCnaive}
\alias{HCnaive}
\title{Hierarchical clustering using naive Bayes}
\usage{
HCnaive(slidedata, markers, probthr = 0.8)
}
\arguments{
\item{slidedata}{data frame of the markers for each cell.}

\item{markers}{data frame indicating the hierarchical structure of the cluster types.}

\item{probthr}{Threshold to define marker positive cells}
}
\description{
Old script that combines separate clustering algorithms together in a hierarchical approach. Should be revamped before use.
}
