% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadImagingData.R
\name{loadImagingData}
\alias{loadImagingData}
\title{Loads the images for a set of slide regions}
\usage{
loadImagingData(
  slideRegionList,
  markers,
  adjustedImageDir = "/media/disk2/atlas_mxif/BALANCED",
  maskdir = "/media/disk2/atlas_mxif/BALANCED"
)
}
\arguments{
\item{slideRegionList}{A character vector of slide_regions to load in. Slide and region should be separated by an underscore.}

\item{markers}{qcname .csv file with qc variables}

\item{adjustedImageDir}{directory of the AF adjusted images.}

\item{maskdir}{directory of where the mask files are stored.}
}
\value{
Returns data frame with image directories.
}
\description{
Just a convenience function that gets directories to all the imaging data on the server.
}
