% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadQuantifiedPCA.R
\name{loadQuantifiedPCA}
\alias{loadQuantifiedPCA}
\title{Loads the quantifed cell level PCA atlas data into memory}
\usage{
loadQuantifiedPCA(
 
    dataname = "/media/disk2/atlas_mxif/data/colon map batch 1-3 position corrected w immune labels.rds",
  qcname = "/media/disk2/atlas_mxif/data/Colon MAP QC.csv"
)
}
\arguments{
\item{dataname}{A data frame with the cell level data}

\item{qcname}{.csv file with qc variables}
}
\value{
Returns data frame with the data merged.
}
\description{
Just a convenience function so that we're all using the same data set
}
